Update JavaDoc and moved files and folders around to make it organized by layer and also by feature
