+++ src/main/java/Main.java
@@ -1,10 +1,9 @@
-import chat.*;
-import chat_use_case.*;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import entities.MessageFactory;
 
-import java.io.FileWriter;
+import controller_presenter_gateway.chat_controller_presenter_gateway.*;
+import entities.MessageFactory;
+import ui.ChatView;
+import use_cases.chat_use_cases.*;
+
 import java.io.IOException;
 import java.util.*;
 import javax.swing.*;
@@ -14,16 +13,22 @@ import java.util.List;
 public class Main {
 
     public static void main(String[] args) throws IOException {
-//        MessageRepository messages = new MessageRepository("message.json");
-//        MessageRepoRequestModel message = new MessageRepoRequestModel(0, "hello", 1, 2, new Date(), new Date(), false, false, false , -1);
-//        messages.save(message);
-//        messages.delete(2);
 
+        // this is to set up something to demo, comment out later
         ChatRepository chat = new ChatRepository("chat.json");
         List<Integer> ids = new ArrayList<>();
-//        ids.add(0);
-//
+        ids.add(0);
+        ids.add(1);
+
         chat.save(new ChatRepoRequestModel(0, ids, false));
+
+        MessageRepoGateway messageGateway = new MessageRepository("message.json");
+        MessageRepoRequestModel message = new MessageRepoRequestModel(0, "hello", 0, 1, new Date(), new Date(), false, false, false , -1);
+        messageGateway.save(message);
+        MessageRepoRequestModel message1 = new MessageRepoRequestModel(1, "hello there", 1, 0, new Date(), new Date(), false, false, false , -1);
+        messageGateway.save(message1);
+
+
         JFrame application = new JFrame("Chat");
         application.setSize(600, 600);
         application.setPreferredSize(new Dimension(600, 600));
@@ -35,32 +40,23 @@ public class Main {
         MessageRepoGateway messageRepoGateway = new MessageRepository("message.json");
         ChatRepoGateway chatRepoGateway = new ChatRepository("chat.json");
         ChatOutputBoundary presenter = new ChatPresenter(view, chatRepoGateway, messageRepoGateway);
-        Map<Integer, ChatRepoRequestModel> chats = chatRepoGateway.getAllChats();
         MessageFactory factory = new MessageFactory();
         DeleteMessageInputBoundary delete = new DeleteMessage(presenter, messageRepoGateway);
         SendMessageInputBoundary send = new SendMessage(factory, presenter, messageRepoGateway, chatRepoGateway);
         EditMessageInputBoundary edit = new EditMessage(presenter, messageRepoGateway);
         ChatController controller = new ChatController(delete, edit, send);
         view.setController(controller);
-//        List<Integer> ids = new ArrayList<>();
-//        ids.add(0);
-//        ids.add(1);
-//        ids.add(2);
-////        ids.add(3);
-////        ids.add(4);
-        List<MessageRepoRequestModel> messages = messageRepoGateway.getMessages(chats.get(1).getMessageIds());
-        view.setChatId(1);
-        view.setUserId(1);
-        view.setOtherUser(2);
-        view.addMessages(messages);
+
+        JPanel panel = new test(presenter);
+
+        presenter.openChat(0, 0, 1);
 
 
         screens.add(view, "Chat");
+        screens.add(panel, "hello");
         cardLayout.show(screens, "Chat");
         application.pack();
         application.setVisible(true);
-
-
