--- src/main/java/chat_use_case/DeleteMessageInputBoundary.java
@@ -1,8 +1,0 @@
-package chat_use_case;
-
-import java.io.IOException;
-
-public interface DeleteMessageInputBoundary {
-
-    void delete(int messageId) throws IOException;
-}
