--- src/main/java/chat_use_case/SendMessageInputBoundary.java
@@ -1,11 +1,0 @@
-package chat_use_case;
-
-import chat.ChatRequestModel;
-
-import java.io.IOException;
-
-public interface SendMessageInputBoundary {
-    void send(ChatRequestModel requestModel) throws IOException;
-
-    void reply(ChatRequestModel requestModel, int messageId) throws IOException;
-}
