+++ src/main/java/ui/ChatViewInterface.java
@@ -1,4 +1,7 @@
-package chat;
+package ui;
+
+import controller_presenter_gateway.chat_controller_presenter_gateway.ChatResponseModel;
+import controller_presenter_gateway.chat_controller_presenter_gateway.MessageRepoRequestModel;
 
 import java.util.List;
 
@@ -13,5 +16,7 @@ public interface ChatViewInterface {
     void addReply(ChatResponseModel responseModel, int replyToId);
 
     void editMessage(int messageId, String content);
+
+    void failView(String message);
 
 }
