+++ src/main/java/use_cases/chat_use_cases/DeleteMessageInputBoundary.java
@@ -1,0 +1,6 @@
+package use_cases.chat_use_cases;
+
+public interface DeleteMessageInputBoundary {
+
+    void delete(int messageId);
+}
