+++ src/main/java/use_cases/chat_use_cases/EditMessage.java
@@ -1,7 +1,7 @@
-package chat_use_case;
+package use_cases.chat_use_cases;
 
-import chat.ChatOutputBoundary;
-import chat.MessageRepoGateway;
+import controller_presenter_gateway.chat_controller_presenter_gateway.ChatOutputBoundary;
+import controller_presenter_gateway.chat_controller_presenter_gateway.MessageRepoGateway;
 
 import java.io.IOException;
 
@@ -11,6 +11,12 @@ public class EditMessage implements EditMessageInputBoundary {
 
     private final MessageRepoGateway messageRepoGateway;
 
+    /**
+     * Creates the use case with the given interfaces
+     *
+     * @param chatOutputBoundary the output boundary
+     * @param messageRepoGateway the message repository gateway
+     */
     public EditMessage(ChatOutputBoundary chatOutputBoundary, MessageRepoGateway messageRepoGateway) {
         this.chatOutputBoundary = chatOutputBoundary;
         this.messageRepoGateway = messageRepoGateway;
@@ -24,9 +30,12 @@ public class EditMessage implements EditMessageInputBoundary {
      * @param content the message
      */
     @Override
-    public void edit(int messageId, String content) throws IOException {
-        messageRepoGateway.edit(messageId, content);
-
-        chatOutputBoundary.editMessage(messageId, content);
+    public void edit(int messageId, String content) {
+        try{
+            messageRepoGateway.edit(messageId, content);
+            chatOutputBoundary.editMessage(messageId, content);
+        } catch (IOException e) {
+            chatOutputBoundary.failView("Message failed to edit");
+        }
     }
 }
