+++ src/main/java/use_cases/chat_use_cases/EditMessageInputBoundary.java
@@ -1,0 +1,5 @@
+package use_cases.chat_use_cases;
+
+public interface EditMessageInputBoundary {
+    void edit(int messageId, String content);
+}
