+++ src/test/java/chat_use_cases/EditMessageTest.java
@@ -1,11 +1,13 @@
-package chat_use_case;
+package chat_use_cases;
 
-import chat.*;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import controller_presenter_gateway.chat_controller_presenter_gateway.*;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import use_cases.chat_use_cases.EditMessage;
+import use_cases.chat_use_cases.EditMessageInputBoundary;
 
 import java.io.FileWriter;
 import java.io.IOException;
