+++ src/test/java/chat_use_cases/SendMessageTest.java
@@ -1,12 +1,14 @@
-package chat_use_case;
+package chat_use_cases;
 
-import chat.*;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import controller_presenter_gateway.chat_controller_presenter_gateway.*;
 import entities.MessageFactory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import use_cases.chat_use_cases.SendMessage;
+import use_cases.chat_use_cases.SendMessageInputBoundary;
 
 import java.io.FileWriter;
 import java.io.IOException;
